fix: path types
