+++ src/pages/catalog-product/api/request-to-get-all-products.ts
@@ -1,6 +1,6 @@
 import { API_HOST, PROJECT_KEY } from "../../../project-config";
+import type { DataProduct } from "../../../shared";
 import { getTokenFromCookie, TOKEN_NAMES } from "../../../shared";
-import type { DataProduct } from "./types";
 
 export async function getAllProducts(): Promise<DataProduct> {
   let allProducts: DataProduct;
